overwriting is optional

+++ DMES.py
@@ -13,18 +13,33 @@ import sys
 from ElapseTime import ElapseTime
 from SpringSupport import SpringSupport
 
+def str2bool(v):
+    if v.lower() in ('yes', 'true', 't', 'y', '1'):
+        return True
+    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
+        return False
+    else:
+        raise argparse.ArgumentTypeError('Boolean value expected.')
+
 # CLI Generation and Parsing
 cliParser = argparse.ArgumentParser()
 cliParser.add_argument('-u', required = False, default = None, help = 'Gemma username')
 cliParser.add_argument('-p', required = False, default = None, help = 'Gemma password')
+cliParser.add_argument('-o', required = False, default = 'Output/', help = 'Output directory')
 cliParser.add_argument('-i', required = True, default = None, help = 'File path for requested experiments')
 cliParser.add_argument('-n', required = False, default = False, action = 'store_true', help = 'Short name flag')
+cliParser.add_argument('--ow', dest='overwrite', action='store_true')
+cliParser.add_argument('--noow', dest='overwrite', action='store_false')
+cliParser.set_defaults(overwrite=True)
+
+
 cliOpts = cliParser.parse_args()
 
 # Declaring Global Flags and Paths
 valUser = None
 valPassword = None
-outputFolder = 'Output/'
+outputFolder = cliOpts.o + '/'
+overwrite = cliOpts.overwrite
 globalTimer = ElapseTime()
 
 if not os.path.exists(outputFolder):
@@ -50,8 +65,17 @@ with open(cliOpts.i, mode = 'rt') as inputFile:
 		tempExpHolder.append(everyLine.strip('\n'))
 
 # Filter Non-existent Experiments
+
+
+
+print(cliOpts.n)
 if cliOpts.n:
 	for eeName in tempExpHolder:
+		tempPath = '{0}{1}.txt'.format(outputFolder, experiment.shortName)
+		# check to see if the file exists
+		if (os.path.lexists(tempPath) and not(overwrite)):
+			continue
+		
 		experiment = experimentService.findByShortName(eeName)
 		if (experiment != None):
 			experimentHolder.append(experiment)
@@ -60,6 +84,11 @@ if cliOpts.n:
 else:
 	for eeID in tempExpHolder:
 		experiment = experimentService.load(long(eeID))
+		# check to see if the file exists
+		tempPath = '{0}{1}.txt'.format(outputFolder, experiment.shortName)
+		if (os.path.lexists(tempPath) and not(overwrite)):
+			continue
+		
 		if (experiment != None):
 			experimentHolder.append(experiment)
 		else:
@@ -73,7 +102,11 @@ for eeIndex, experiment in enumerate(experimentHolder, start = 1):
 	
 	tempPath = '{0}{1}.txt'.format(outputFolder, experiment.shortName)
 	if (os.path.lexists(tempPath)):
-		os.remove(tempPath)
+		if(overwrite):
+			os.remove(tempPath)
+		else:
+			continue
+	
 	
 	try:
 		outFile = edfService.writeDataFile(experiment, False, tempPath, False)
